requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

PyYAML == 6.0
matplotlib == 3.1.2
